refactor(infra): remove v2/CubeType, v2/PotentialGrade, use core enums
