+++ module-app/src/main/java/maple/expectation/application/service/cube/AbstractCubeDecoratorV2.java
@@ -5,7 +5,7 @@ import maple.expectation.application.service.calculator.EnhanceDecorator;
 import maple.expectation.application.service.calculator.ExpectationCalculator;
 import maple.expectation.application.service.cube.policy.CubeCostPolicy;
 import maple.expectation.core.dto.cube.CubeCalculationInput;
-import maple.expectation.domain.v2.CubeType;
+import maple.expectation.core.domain.model.CubeType;
 
 /**
  * V2-specific abstract cube decorator using Long type.
