+++ module-infra/src/main/java/maple/expectation/application/service/cube/AbstractCubeDecorator.java
@@ -4,7 +4,7 @@ import java.math.RoundingMode;
 import java.util.Optional;
 import maple.expectation.application.service.cube.policy.CubeCostPolicy;
 import maple.expectation.core.dto.cube.CubeCalculationInput;
-import maple.expectation.domain.v2.CubeType;
+import maple.expectation.core.domain.model.CubeType;
 
 /**
  * Abstract template for Cube decorators that eliminates duplication between V2 (long) and V4
