+++ module-infra/src/main/java/maple/expectation/application/service/cube/AbstractCubeDecoratorV4.java
@@ -6,7 +6,7 @@ import maple.expectation.application.service.calculator.v4.EquipmentExpectationC
 import maple.expectation.application.service.calculator.v4.EquipmentExpectationCalculator.CostBreakdown;
 import maple.expectation.application.service.cube.policy.CubeCostPolicy;
 import maple.expectation.core.dto.cube.CubeCalculationInput;
-import maple.expectation.domain.v2.CubeType;
+import maple.expectation.core.domain.model.CubeType;
 
 /**
  * V4-specific abstract cube decorator using Double type for performance.
