+++ module-infra/src/main/kotlin/maple/expectation/application/service/calculator/v4/impl/RedCubeDecoratorV4.kt
@@ -5,7 +5,7 @@ import maple.expectation.application.service.calculator.v4.EquipmentExpectationC
 import maple.expectation.application.service.cube.AbstractCubeDecoratorV4
 import maple.expectation.application.service.cube.CubeTrialsProvider
 import maple.expectation.application.service.cube.policy.CubeCostPolicy
-import maple.expectation.domain.v2.CubeType
+import maple.expectation.core.domain.model.CubeType
 import maple.expectation.core.dto.cube.CubeCalculationInput
 
 /**
