--- module-infra/src/main/kotlin/maple/expectation/domain/v2/CubeType.kt
@@ -1,27 +1,0 @@
-package maple.expectation.domain.v2
-
-/**
- * 큐브 종류 Enum
- */
-enum class CubeType(val description: String) {
-    BLACK("블랙큐브"),
-    RED("레드큐브"),
-    ADDITIONAL("에디셔널큐브"),
-    ;
-
-    /** Convert to core CubeType */
-    fun toCore(): maple.expectation.core.domain.model.CubeType = when (this) {
-        BLACK -> maple.expectation.core.domain.model.CubeType.BLACK
-        RED -> maple.expectation.core.domain.model.CubeType.RED
-        ADDITIONAL -> maple.expectation.core.domain.model.CubeType.ADDITIONAL
-    }
-
-    companion object {
-        @JvmStatic
-        fun fromCore(core: maple.expectation.core.domain.model.CubeType): CubeType = when (core) {
-            maple.expectation.core.domain.model.CubeType.BLACK -> BLACK
-            maple.expectation.core.domain.model.CubeType.RED -> RED
-            maple.expectation.core.domain.model.CubeType.ADDITIONAL -> ADDITIONAL
-        }
-    }
-}
