+++ module-infra/src/main/kotlin/maple/expectation/infrastructure/persistence/repository/CubeProbabilityRepositoryImpl.kt
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.dataformat.csv.CsvSchema
 import jakarta.annotation.PostConstruct
 import maple.expectation.domain.repository.CubeProbabilityRepository
 import maple.expectation.domain.v2.CubeProbability
-import maple.expectation.domain.v2.CubeType
+import maple.expectation.core.domain.model.CubeType
 import maple.expectation.error.exception.CubeDataInitializationException
 import maple.expectation.infrastructure.executor.LogicExecutor
 import maple.expectation.infrastructure.executor.TaskContext
